+++ RandomSkunk.Results.AspNetCore/ErrorExtensions.cs
@@ -10,6 +10,35 @@ public static class ErrorExtensions
         errorCode => Math.Abs(errorCode) % 1000;
 
     /// <summary>
+    /// Gets an <see cref="IActionResult"/> that represents the source <see cref="Error"/>.
+    /// </summary>
+    /// <param name="sourceError">The source error.</param>
+    /// <param name="type">A URI reference [RFC3986] that identifies the problem type. This specification encourages that, when
+    ///     dereferenced, it provide human-readable documentation for the problem type (e.g., using HTML
+    ///     [W3C.REC-html5-20141028]). When this member is not present, its value is assumed to be "about:blank".</param>
+    /// <param name="instance">A URI reference that identifies the specific occurrence of the problem. It may or may not yield
+    ///     further information if dereferenced.</param>
+    /// <param name="getHttpStatusCode">An optional function that is used to get an HTTP status code from an
+    ///     <see cref="Error.ErrorCode"/>. If <see langword="null"/> or not provided, then the following function is used:
+    ///     <code>errorCode => Math.Abs(errorCode) % 1000</code>
+    ///     This function discards the sign of the number and all but the last three digits of the number are used. For example,
+    ///     passing -123456 returns 456.</param>
+    /// <returns>An <see cref="ObjectResult"/> for a <see cref="ProblemDetails"/> describing the error.</returns>
+    public static IActionResult GetActionResult(
+        this Error sourceError,
+        string? type = null,
+        string? instance = null,
+        Func<int, int>? getHttpStatusCode = null)
+    {
+        var httpStatusCode = sourceError.GetHttpStatusCode(getHttpStatusCode);
+
+        return new ObjectResult(sourceError.GetProblemDetails(type, instance, getHttpStatusCode))
+        {
+            StatusCode = httpStatusCode ?? ErrorCodes.InternalServerError,
+        };
+    }
+
+    /// <summary>
     /// Gets a <see cref="ProblemDetails"/> object that is equavalent to the error object.
     /// </summary>
     /// <param name="sourceError">The <see cref="Error"/> to create a <see cref="ProblemDetails"/> from.</param>
@@ -21,7 +50,8 @@ public static class ErrorExtensions
     /// <param name="getHttpStatusCode">An optional function that is used to get an HTTP status code from the
     ///     <see cref="Error.ErrorCode"/>. If <see langword="null"/> or not provided, then the following function is used:
     ///     <code>errorCode => Math.Abs(errorCode) % 1000</code>
-    /// </param>
+    ///     This function discards the sign of the number and all but the last three digits of the number are used. For example,
+    ///     passing -123456 returns 456.</param>
     /// <returns>The equivalent problem details object.</returns>
     /// <exception cref="ArgumentNullException">If <paramref name="sourceError"/> is <see langword="null"/>.</exception>
     public static ProblemDetails GetProblemDetails(
@@ -64,6 +94,8 @@ public static class ErrorExtensions
     /// <remarks>
     /// This method maps an error code to an HTTP status code according to the following function:
     /// <code>errorCode => Math.Abs(errorCode) % 1000</code>
+    /// This function discards the sign of the number and all but the last three digits of the number are used. For example,
+    /// passing -123456 returns 456.
     /// </remarks>
     /// <param name="sourceError">The <see cref="Error"/> to get an HTTP status code from.</param>
     /// <returns>The HTTP status code.</returns>